examples/requirements.txt
python-dotenv
weclappy
pytest
pytest-cov